src/main/java/com/ctci/ch1/Ch1_1.java
package com.ctci.ch1;

import java.util.HashSet;
import java.util.Set;

public class Ch1_1 {
    public Boolean hasUnique(String s) {
        Set<Character> set = new HashSet<>();

        for (char c : s.toCharArray()) {
            if (set.contains(c)) {
                return false;
            } else {
                set.add(c);
            }
        }
        return true;
    }

    public Boolean checkUniqueBruteForce(String s) {
        for (int i = 0; i < s.length(); i++) {
            for (int j = i + 1; j < s.length(); j++) {
                if (s.charAt(i) == s.charAt(j)) {
                    return false;
                }
            }
        }
        return true;
    }
}
